Emoji e2e tests to typescript migration
