+++ e2e/cypress/tests/integration/emoji/custom_emoji_2_1_spec.ts
