+++ e2e/cypress/tests/integration/emoji/recently_used_emoji_1_spec.ts
