+++ e2e/cypress/tests/integration/emoji/recently_used_emoji_2_spec.ts
