+++ e2e/cypress/tests/integration/emoji/sorted_emojis_spec.ts
